tools/requirements.txt
beautifulsoup4
lxml
youtube-transcript-api
openai
tiktoken
markdown-it-py
pyyaml
selenium
webdriver-manager 